Add 'try_extract_send_packet_event' to relayer-cosmos methods module
