+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -5,6 +5,7 @@ pub mod connection;
 pub mod consensus_state;
 pub mod create_client;
 pub mod encode;
+pub mod event;
 pub mod packet;
 pub mod runtime;
 pub mod unreceived_packet;
